Abstracted SearchBox (DESC)
Abstracted SearchBox away so it can be re-used in multiple components and removed the anti-pattern of passing the setter down to the SearchBox component, instead opting to define a callback that gets called with the onChange prop of the TextInput.
